Trike gravity, sound, and jump turn improvements
Fixed stupid gravity issue. When crushing gravity due to ground contact, set it to -0.4, which seems to be the minimum that the CharacterController respects for on-ground detection.

Fallout of above is that trike audio didn't know how to stop due to speed decrease, so I changed the rolling audio manipulation to be based on curSpeed/maxSpeed interpolation.

Gravity fix resolves jump turn issue, and I realized I was not setting the jump turn speed high enough.